Move allocation removal code into a plugin and fix and refactor it
