+++ coldfront/components/site/templates/allocation/allocation_detail.html
@@ -41,15 +41,23 @@ Allocation Detail
           <h3><i class="fas fa-list" aria-hidden="true"></i> Allocation Information</h3>
         </div>
         <div class="col">
-          {% if allocation.is_changeable %}
+          {% if allocation.is_changeable and allocation.status.name == Active %}
             <a class="btn btn-primary" href="{% url 'allocation-change' allocation.pk %}" role="button" style="float: right;">
               Request Change
             </a>
           {% endif %}
-          {% if allocation.status.name in 'Active, Billing Information Submitted' and allocation.get_parent_resource.requires_payment %}
-            <a class="btn btn-primary mr-1" href="{% url 'allocation-remove' allocation.pk %}" role="button" style="float: right;">
-              Remove Access
-            </a>
+          {% if allocation.status.name in 'Active, Billing Information Submitted' and allocation.get_parent_resource.name in ALLOCATION_REMOVAL_REQUESTS_ALLOWED %}
+            {% load allocation_removal_requests_tags %}
+            {% has_admin_perm allocation request.user 'allocation_removal_requests.can_remove_allocations' as has_perm %}
+            {% if has_perm %}
+              <a class="btn btn-primary mr-1" href="{% url 'allocation_removal_requests:allocation-remove' allocation.pk %}" role="button" style="float: right;">
+                Remove Allocation
+              </a>
+            {% else %}
+              <a class="btn btn-primary mr-1" href="{% url 'allocation_removal_requests:allocation-removal-request' allocation.pk %}" role="button" style="float: right;">
+                Request Removal
+              </a>
+            {% endif %}
           {% endif %}
         </div>
       </div>
