+++ coldfront/components/site/templates/common/navbar_admin.html
@@ -9,7 +9,7 @@
     <a id="navbar-project-reviews" class="dropdown-item" href="{% url 'project-review-list' %}">Project Reviews/Requests</a>
     <a id="navbar-allocation-requests" class="dropdown-item" href="{% url 'allocation-request-list' %}">
       Allocation Requests</a>
-    <a id="navbar-allcoation-remove-requests" class="dropdown-item" href="{% url 'allocation-removal-request-list' %}">
+    <a id="navbar-allcoation-remove-requests" class="dropdown-item" href="{% url 'allocation_removal_requests:allocation-removal-request-list' %}">
       Allocation Removal Requests</a>
     <a id="navbar-allocation-change-request" class="dropdown-item" href="{% url 'allocation-change-list' %}">
       Allocation Change Requests</a>
