+++ coldfront/config/settings.py
@@ -36,6 +36,7 @@ plugin_configs = {
     'PLUGIN_COLDFRONT_CUSTOM_RESOURCES': 'plugins/coldfront_custom_resources.py',
     'PLUGIN_CUSTOMIZABLE_FORMS': 'plugins/customizable_forms.py',
     'PLUGIN_PI_SEARCH': 'plugins/pi_search.py',
+    'PLUGIN_ALLOCATION_REMOVAL_REQUESTS':'plugins/allocation_removal_requests.py'
 }
 
 # This allows plugins to be enabled via environment variables. Can alternatively
