+++ coldfront/config/urls.py
@@ -6,6 +6,7 @@ from django.contrib import admin
 from django.urls import include, path
 from django.views.generic import TemplateView
 from django.conf.urls.static import static
+from debug_toolbar.toolbar import debug_toolbar_urls
 
 import coldfront.core.portal.views as portal_views
 
@@ -28,7 +29,10 @@ urlpatterns = [
     path('grant/', include('coldfront.core.grant.urls')),
     path('publication/', include('coldfront.core.publication.urls')),
     path('research-output/', include('coldfront.core.research_output.urls')), 
-] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
+] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT) + debug_toolbar_urls()
+
+if 'coldfront.plugins.allocation_removal_requests' in settings.INSTALLED_APPS:
+    urlpatterns.append(path('allocation_removal_requests/', include('coldfront.plugins.allocation_removal_requests.urls')))
 
 if 'coldfront.plugins.pi_search' in settings.INSTALLED_APPS:
     urlpatterns.append(path('pi_search_function/', include('coldfront.plugins.pi_search.urls')))
