+++ coldfront/core/allocation/admin.py
@@ -22,8 +22,6 @@ from coldfront.core.allocation.models import (Allocation, AllocationAccount,
                                               AllocationInvoice,
                                               AllocationAdminAction,
                                               AttributeType,
-                                              AllocationRemovalRequest,
-                                              AllocationRemovalStatusChoice,
                                               AllocationUserRoleChoice,)
 
 from coldfront.core.resource.models import Resource
@@ -556,33 +554,3 @@ class AllocationAdminActionAdmin(ReviewGroupFilteredResourceQueryset):
             return super().get_readonly_fields(request)
         else:
             return self.readonly_fields_change
-
-
-@admin.register(AllocationRemovalRequest)
-class AllocationRemovalRequestAdmin(ReviewGroupFilteredResourceQueryset):
-    list_display = ('pk', 'allocation_pk', 'project_pi', 'requestor', 'allocation_prior_status',
-                    'resource', 'status')
-    readonly_fields = ('project_pi', 'requestor', 'allocation_prior_status', 'allocation')
-    list_filter = (
-        'status',
-        'allocation__resources',
-        'allocation_prior_status'
-    )
-    search_fields = (
-        'requestor__username',
-        'requestor__first_name',
-        'requestor__last_name',
-    )
-
-    def resource(self, obj):
-        allocation_obj = obj.allocation
-        return allocation_obj.get_parent_resource.name
-    
-    def allocation_pk(self, obj):
-        allocation_obj = obj.allocation
-        return allocation_obj.pk
-
-
-@admin.register(AllocationRemovalStatusChoice)
-class AllocationRemovalStatusChoiceAdmin(admin.ModelAdmin):
-    list_display = ('pk', 'name')
