+++ coldfront/core/allocation/signals.py
@@ -17,8 +17,6 @@ allocation_change_approved = django.dispatch.Signal()
 
 allocation_expire = django.dispatch.Signal()
     # providing_args=["allocation_pk"])
-allocation_remove= django.dispatch.Signal()
-    #providing_args=["allocation_pk"])
 allocation_change_user_role = django.dispatch.Signal()
     # providing_args=["allocation_user_pk"])
 visit_allocation_detail = django.dispatch.Signal()
