+++ coldfront/plugins/allocation_removal_requests/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'coldfront.plugins.allocation_removal_requests.apps.AllocationRemovalRequestsConfig'
