+++ coldfront/plugins/allocation_removal_requests/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AllocationRemovalRequestsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'coldfront.plugins.allocation_removal_requests'
