+++ coldfront/plugins/allocation_removal_requests/signals.py
@@ -1,0 +1,3 @@
+import django.dispatch
+
+allocation_remove = django.dispatch.Signal()
