+++ coldfront/plugins/allocation_removal_requests/templates/allocation_removal_requests/allocation_remove.html
@@ -10,6 +10,7 @@ Remove Allocation
 <h2>Remove Allocation</h2>
 {% if is_admin %}
   <div class="alert alert-warning" role="alert">
+    <i class="fa fa-exclamation-circle"></i>
     This allocation will be removed immediately.
   </div>
 {% endif %}
@@ -35,13 +36,10 @@ Remove Allocation
 
 <div class="alert alert-danger" role="alert">
   <form method="post">{% csrf_token %}
-    <p>Are you sure you want to remove access to this allocation?</p>
+    <p>Are you sure you want to remove access to this allocation?
+      Doing so will cause everyone in this allocation to lose access to its resource</p>
     <input class="btn btn-danger" type="submit" value="Yes" />
-    {% if from_project %}
-      <a class="btn btn-secondary" href="{% url 'project-detail' allocation.project.pk %}" role="button">Cancel</a>
-    {% else %}
-      <a class="btn btn-secondary" href="{% url 'allocation-detail' allocation.pk %}" role="button">Cancel</a>
-    {% endif %}
+    <a class="btn btn-secondary" href="{% url 'allocation-detail' allocation.pk %}" role="button">Cancel</a>
   </form>
 </div>
