+++ coldfront/plugins/allocation_removal_requests/templates/allocation_removal_requests/allocation_removed.txt
@@ -4,5 +4,7 @@ Your allocation for {{resource}} has been removed. You no longer have access to 
 
 To view your allocation's information, please go to {{url}}
 
+If you are not the PI or a manager under this project, you are receiving this notice as a courtesy.
+
 Thank you,
 {{signature}}
